Moved commit + reveal messages to transactions.py, added FactoidTransaction message
